pyproject.toml
[project]
name = "home-hub"
version = "2.0.0"
description = "Bonifacic Home Management App"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic-settings>=2.0.0",
    "bcrypt>=4.0.0",
    "slowapi>=0.1.9",
    "httpx>=0.27.0",
    "anthropic>=0.40.0",
    "caldav>=1.4.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dev = "uvicorn backend.main:app --port 8400 --reload"

[tool.ruff]
line-length = 120
